Add tracking of last updated time

+++ TableService.cs
@@ -9,20 +9,17 @@ namespace SchoolHelpdesk;
 public static class TableService
 {
   private static TableClient ticketsClient;
-  private static TableClient commentsClient;
   private static int latestTicketId = -1;
 
   public static void Configure(string connectionString)
   {
     ticketsClient = new TableServiceClient(connectionString).GetTableClient("tickets");
-    commentsClient = new TableServiceClient(connectionString).GetTableClient("comments");
   }
 
   public static async Task WarmUpAsync()
   {
     var nonExistentKey = "warmup";
     await ticketsClient.QueryAsync<TableEntity>(o => o.PartitionKey == nonExistentKey).ToListAsync();
-    await commentsClient.QueryAsync<TableEntity>(o => o.PartitionKey == nonExistentKey).ToListAsync();
   }
 
   public static async Task<TicketEntity> GetTicketAsync(string assigneeEmail, int id)
@@ -38,17 +35,21 @@ public static class TableService
     return tickets.Count == 1 ? tickets[0] : null;
   }
 
-  public static async Task<List<TicketEntity>> GetAllTicketsAsync()
+  public static async Task<List<TicketEntity>> GetAllTicketsAsync(DateTime? after = null)
   {
-    var tickets = await ticketsClient.QueryAsync<TicketEntity>().ToListAsync();
-    return tickets.OrderByDescending(t => t.Timestamp).ToList();
+    var tickets = after is null
+      ? await ticketsClient.QueryAsync<TicketEntity>().ToListAsync()
+      : await ticketsClient.QueryAsync<TicketEntity>(o => o.LastUpdated > after.Value || !o.IsClosed).ToListAsync();
+    return tickets.OrderByDescending(t => t.LastUpdated).ToList();
   }
 
-  public static async Task<List<TicketEntity>> GetTicketsByAssigneeAsync(string assigneeEmail)
+  public static async Task<List<TicketEntity>> GetTicketsByAssigneeAsync(string assigneeEmail, DateTime? after = null)
   {
     ArgumentNullException.ThrowIfNull(assigneeEmail);
-    var tickets = await ticketsClient.QueryAsync<TicketEntity>(o => o.PartitionKey == assigneeEmail).ToListAsync();
-    return tickets.OrderByDescending(t => t.Timestamp).ToList();
+    var tickets = after is null
+      ? await ticketsClient.QueryAsync<TicketEntity>(o => o.PartitionKey == assigneeEmail).ToListAsync()
+      : await ticketsClient.QueryAsync<TicketEntity>(o => o.PartitionKey == assigneeEmail && (o.LastUpdated > after.Value || !o.IsClosed)).ToListAsync();
+    return tickets.OrderByDescending(t => t.LastUpdated).ToList();
   }
 
   public static async Task<bool> TicketExistsAsync(string assigneeEmail, int id)
@@ -65,9 +66,11 @@ public static class TableService
 
     var id = Interlocked.Increment(ref latestTicketId);
     ticket.Created = DateTime.UtcNow;
+    ticket.LastUpdated = ticket.Created;
     ticket.WaitingSince = ticket.Created;
     ticket.RowKey = id.ToRowKey();
-    await ticketsClient.AddEntityAsync(ticket);
+    var response = await ticketsClient.AddEntityAsync(ticket);
+    ticket.ETag = response.Headers.ETag.Value;
     return id;
   }
 
@@ -75,12 +78,15 @@ public static class TableService
   {
     ArgumentNullException.ThrowIfNull(assigneeEmail);
     ArgumentNullException.ThrowIfNull(newAssigneeEmail);
+    if (assigneeEmail == newAssigneeEmail) throw new ArgumentException("New assignee email must be different from current assignee email.");
 
     var rowKey = id.ToRowKey();
     var ticket = await ticketsClient.GetEntityAsync<TicketEntity>(assigneeEmail, rowKey);
     ticket.Value.PartitionKey = newAssigneeEmail;
     ticket.Value.AssigneeName = newAssigneeName;
-    await ticketsClient.AddEntityAsync(ticket.Value);
+    ticket.Value.LastUpdated = DateTime.UtcNow;
+    var response = await ticketsClient.AddEntityAsync(ticket.Value);
+    ticket.Value.ETag = response.Headers.ETag.Value;
     await ticketsClient.DeleteEntityAsync(assigneeEmail, rowKey);
     return ticket.Value;
   }
@@ -91,7 +97,7 @@ public static class TableService
     ArgumentNullException.ThrowIfNull(newTitle);
     var ticket = await ticketsClient.GetEntityAsync<TicketEntity>(assigneeEmail, id.ToRowKey());
     ticket.Value.Title = newTitle;
-    await ticketsClient.UpdateEntityAsync(ticket.Value, ETag.All, TableUpdateMode.Replace);
+    await ticketsClient.UpdateEntityAsync(ticket.Value, ticket.Value.ETag, TableUpdateMode.Replace);
   }
 
   public static async Task CloseTicketAsync(string assigneeEmail, int id, bool isClosed)
@@ -99,8 +105,10 @@ public static class TableService
     ArgumentNullException.ThrowIfNull(assigneeEmail);
     var ticket = await ticketsClient.GetEntityAsync<TicketEntity>(assigneeEmail, id.ToRowKey());
     ticket.Value.IsClosed = isClosed;
-    await ticketsClient.UpdateEntityAsync(ticket.Value, ETag.All, TableUpdateMode.Replace);
+    ticket.Value.LastUpdated = DateTime.UtcNow;
+    await ticketsClient.UpdateEntityAsync(ticket.Value, ticket.Value.ETag, TableUpdateMode.Replace);
   }
+
   public static async Task ChangeTicketStudentAsync(TicketEntity ticket, Student student)
   {
     ArgumentNullException.ThrowIfNull(ticket);
@@ -108,7 +116,8 @@ public static class TableService
     ticket.StudentLastName = student?.LastName;
     ticket.TutorGroup = student?.TutorGroup;
     ticket.ParentRelationship = student?.ParentRelationship;
-    await ticketsClient.UpdateEntityAsync(ticket, ETag.All, TableUpdateMode.Replace);
+    var response = await ticketsClient.UpdateEntityAsync(ticket, ticket.ETag, TableUpdateMode.Replace);
+    ticket.ETag = response.Headers.ETag.Value;
   }
 
   public static async Task ChangeTicketParentAsync(TicketEntity ticket, Parent parent, string relationship)
@@ -119,24 +128,27 @@ public static class TableService
     ticket.ParentEmail = parent.Email;
     ticket.ParentPhone = parent.Phone;
     ticket.ParentRelationship = relationship;
-    await ticketsClient.UpdateEntityAsync(ticket, ETag.All, TableUpdateMode.Replace);
+    var response = await ticketsClient.UpdateEntityAsync(ticket, ticket.ETag, TableUpdateMode.Replace);
+    ticket.ETag = response.Headers.ETag.Value;
   }
 
-  public static async Task ClearLastParentMessageDateAsync(string assigneeEmail, int id)
-  {
-    ArgumentNullException.ThrowIfNull(assigneeEmail);
-    var ticket = await ticketsClient.GetEntityAsync<TicketEntity>(assigneeEmail, id.ToRowKey());
-    ticket.Value.WaitingSince = null;
-    await ticketsClient.UpdateEntityAsync(ticket.Value, ETag.All, TableUpdateMode.Replace);
-  }
-
-  public static async Task SetLastParentMessageDateAsync(TicketEntity ticket)
+  public static async Task SetLastUpdatedAsync(TicketEntity ticket, DateTime lastUpdated, bool cancelWaiting)
   {
     ArgumentNullException.ThrowIfNull(ticket);
-    if (ticket.WaitingSince is not null && !ticket.IsClosed) return;
-    ticket.WaitingSince ??= DateTime.UtcNow;
+    ticket.LastUpdated = lastUpdated;
+    if (cancelWaiting) ticket.WaitingSince = null;
+    var response = await ticketsClient.UpdateEntityAsync(ticket, ticket.ETag, TableUpdateMode.Replace);
+    ticket.ETag = response.Headers.ETag.Value;
+  }
+
+  public static async Task UpdateForNewParentMessageAsync(TicketEntity ticket, DateTime lastUpdated)
+  {
+    ArgumentNullException.ThrowIfNull(ticket);
+    ticket.LastUpdated = lastUpdated;
+    ticket.WaitingSince ??= lastUpdated;
     ticket.IsClosed = false;
-    await ticketsClient.UpdateEntityAsync(ticket, ETag.All, TableUpdateMode.Replace);
+    var response = await ticketsClient.UpdateEntityAsync(ticket, ticket.ETag, TableUpdateMode.Replace);
+    ticket.ETag = response.Headers.ETag.Value;
   }
 
   public static async Task LoadLatestTicketIdAsync()
@@ -165,6 +177,7 @@ public class TicketEntity : ITableEntity
   public bool IsClosed { get; set; }
   public string Title { get; set; }
   public DateTime Created { get; set; }
+  public DateTime LastUpdated { get; set; }
   public DateTime? WaitingSince { get; set; }
   public string StudentFirstName { get; set; }
   public string StudentLastName { get; set; }
